second pass cleanup
