+++ .gitignore
@@ -153,3 +153,55 @@ ENV/
 node_modules/
 mealie/data/debug/last_recipe.json
 *.sqlite
+app_data/backups/.gitkeep
+app_data/backups/dev_sample_data_2021-Jan-12.zip
+app_data/debug/.gitkeep
+app_data/debug/last_recipe.json
+app_data/img/.gitkeep
+app_data/img/banana-bread.jpg
+app_data/img/bon-appetit-s-perfect-pizza.jpg
+app_data/img/braised-beans-and-sardines-with-fennel.jpg
+app_data/img/broccoli-beer-cheese-soup.jpg
+app_data/img/buttery-kimchi-chicken.jpg
+app_data/img/cauliflower-cacciatore.jpg
+app_data/img/chicken-salad-with-citrus-and-chile-oil.jpg
+app_data/img/coffee-hazelnut-biscotti.jpg
+app_data/img/corn-and-crab-beignets-with-yaji-aioli.jpg
+app_data/img/crispy-carrots.jpg
+app_data/img/crispy-rice-with-ginger-citrus-celery-salad.jpg
+app_data/img/crockpot-buffalo-chicken.jpg
+app_data/img/detroit-style-pepperoni-pizza.jpg
+app_data/img/downtown-marinade.jpg
+app_data/img/falafel-hummus-plate.jpg
+app_data/img/five-spice-popcorn-chicken.jpg
+app_data/img/ginger-citrus-cookies.jpg
+app_data/img/green-chile-stew.jpg
+app_data/img/green-seasoning-baked-cod.jpg
+app_data/img/green-spaghetti.jpg
+app_data/img/grilled-mushrooms-and-root-vegetables.jpg
+app_data/img/how-to-make-instant-pot-kerala-vegetable-stew.jpg
+app_data/img/huevos-rancheros-con-rajas-y-champinones.jpg
+app_data/img/instant-pot-chicken-and-potatoes.jpg
+app_data/img/jalapeno-cornbread.jpg
+app_data/img/macadamia-and-brown-butter-blondies.jpg
+app_data/img/marinated-tofu-with-brussels-sprouts-and-farro.jpg
+app_data/img/marranitos-enfiestados.jpg
+app_data/img/mississippi-pot-roast.jpg
+app_data/img/mushroom-risotto.jpg
+app_data/img/new-york-strip.jpg
+app_data/img/nilla-wafer-french-toast.jpg
+app_data/img/one-minute-muffin.jpg
+app_data/img/one-pot-chicken-and-rice.jpg
+app_data/img/pace-pork.jpg
+app_data/img/pasta-with-mushrooms-and-cashew-cream.jpg
+app_data/img/pizzettes.jpg
+app_data/img/pork-steaks.jpg
+app_data/img/roasted-brussels-sprouts.jpg
+app_data/img/roasted-okra.jpg
+app_data/img/salt-vinegar-potatoes.jpg
+app_data/img/shrimp-and-cabbage-curry.jpg
+app_data/img/smashed-carrots.jpg
+app_data/img/tamarind-chicken-thighs-with-collard-greens-salad.jpg
+app_data/img/tequila-beer-and-citrus-cocktail.jpg
+app_data/migration/.gitkeep
+app_data/templates/recipes.md
