+++ Dockerfile.dev
@@ -8,7 +8,6 @@ RUN curl -sSL https://raw.githubusercontent.com/python-poetry/poetry/master/get-
     ln -s /opt/poetry/bin/poetry && \
     poetry config virtualenvs.create false
 
-RUN mkdir /app/
 COPY ./pyproject.toml ./app/poetry.lock* /app/
 
 WORKDIR /app
@@ -16,7 +15,6 @@ WORKDIR /app
 RUN poetry install --no-root
 
 COPY ./mealie /app
-
 
 ENTRYPOINT [ "python" ]
